+++ src/app.vue
@@ -82,53 +82,7 @@
     </div>
 
     <!-- size="lg" -->
-    <b-modal 
-      ref="settings" 
-      centered 
-      hide-header 
-      hide-footer
-      body-class="irisc-modal p-1"
-    >
-
-      <div class="px-5 py-1">
-        <h4>configuration</h4>
-        <div class="mt-3">
-          <!-- simulation speed -->
-          <div>
-            cpu tickrate
-            <span class="float-right">{{ (1000 / delay).toFixed(2) }} tps</span>
-          </div>
-          <b-form-input 
-            :value="1000 / delay"
-            @change="simulatorState.setDelay(1000 / $event)"
-            type="range"
-            min="0.5"
-            max="50"
-            step="0.1"
-          ></b-form-input>
-
-          <!-- memory size -->
-          <div>
-            ram size 
-            <span class="float-right">{{ memory.size }} bytes</span>
-          </div>
-          <b-form-input 
-            :value="memory.sizes.findIndex(e => e === memory.size)"
-            @change="simulatorState.init(memory.sizes[+$event])"
-            type="range"
-            min="0"
-            :max="memory.sizes.length - 1"
-          ></b-form-input>
-
-          <!-- <div>
-            crt effect (epilepsy warning)
-          </div>
-          <b-form-checkbox :checked="settings.crtEffect" @change="settingsState.setCrtEffect($event)" name="crt-effect" switch>
-            {{ settings.crtEffect ? 'on' : 'off' }} 
-          </b-form-checkbox> -->
-        </div>
-      </div>
-    </b-modal>
+    <settings ref="settings"></settings>
 
     <b-modal 
       ref="about" 
@@ -196,7 +150,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { editor, terminal, registers, memory, instruction, tutorial } from "@/vue";
+import { editor, terminal, registers, memory, instruction, tutorial, settings } from "@/vue";
 import { SimulatorState } from "@/simulator";
 import { Interpreter, RuntimeError } from '@/interpreter';
 import { Register, EnvironmentType } from "@/constants"
@@ -218,13 +172,12 @@ export default Vue.extend({
     registers,
     memory,
     instruction,
-    tutorial
+    tutorial, 
+    settings
   },
   data() {
     return {
       env: EnvironmentType.TERMINAL,
-      simulatorState: SimulatorState,
-      settingsState: SettingsState,
 
       dismissTooSmall: false,
       windowSize: 0
@@ -233,9 +186,8 @@ export default Vue.extend({
   computed: {
     currentTick: SimulatorState.currentTick,
 
-    registers: SimulatorState.registers,
-    memory: SimulatorState.memory,
-    breakpoints: SimulatorState.breakpoints,
+    // registers: SimulatorState.registers,
+    // breakpoints: SimulatorState.breakpoints,
     errors: SimulatorState.errors,
     errorSummary: function (): string[] {
       return this.errors.map(e => `${e.constructHelperHTML()}`)
@@ -243,10 +195,7 @@ export default Vue.extend({
     
     running: SimulatorState.running,
     paused: SimulatorState.paused,
-    delay: SimulatorState.delay,
-    step: SimulatorState.step,
-
-    settings: SettingsState.settings,
+    // step: SimulatorState.step,
 
     isTooSmall: function (): boolean {
       return !this.dismissTooSmall && this.windowSize < 1250;
