+++ src/vue/index.ts
@@ -4,5 +4,6 @@ export { default as memory } from "./memory.vue";
 export { default as registers } from "./registers.vue";
 export { default as terminal } from "./terminal.vue";
 export { default as explorer } from "./explorer.vue";
+export { default as settings } from "./settings.vue";
 
 export { default as tutorial } from "./tutorial.vue";
